refactor energy cost calculation functions
