+++ apps/web/src/components/costs/energy-costs-change-last-thirty-days-national-average-card.tsx
@@ -1,3 +1,4 @@
+import { getNationalAverageCost } from "@/components/costs/get-national-average-cost";
 import { calculateCosts } from "@/components/dashboard/energy-cost";
 import { formatNumber } from "@energyleaf/lib";
 import { Card, CardContent, CardHeader, CardTitle } from "@energyleaf/ui/card";
@@ -63,31 +64,6 @@ function EnergyCostsChangeLastThirtyDaysNationalAverage({ userData, energyData }
             </CardContent>
         </Card>
     );
-}
-
-function getNationalAverageCost(userData, days) {
-    const { people, houseType = "apartment", hotWater = "not_electric" } = userData;
-
-    const yearlyCosts = {
-        house: {
-            regular: [1015, 1265, 1520, 1690, 2110],
-            withHotWater: [1140, 1480, 1900, 2155, 2660],
-        },
-        apartment: {
-            regular: [590, 845, 1100, 1225, 1265],
-            withHotWater: [720, 1180, 1520, 1775, 1900],
-        },
-    };
-
-    if (people > 5) return null;
-
-    const costArray = yearlyCosts[houseType]?.[hotWater === "electric" ? "withHotWater" : "regular"];
-    if (!costArray) {
-        return null;
-    }
-
-    const annualCost = costArray[Math.min(people, costArray.length) - 1];
-    return (annualCost / 365) * days;
 }
 
 export default EnergyCostsChangeLastThirtyDaysNationalAverage;
